Update README.md
I really loved using this, but it was tough because I didn't see any examples of how to implement this code, specifically with Node.  I took the liberty to write a small post about how to do this because I couldn't find anything online.  You may want to think about adding it to the Readme because it was a little difficult for me to get this working with Node.  I think it will help people.

Thanks!